+++ Meio.App/Resources/ColorPalette.axaml
