+++ Meio.App/Resources/Styles.axaml
