+++ Meio.App/Services/FileHelper.cs
