+++ Meio.App/Services/ImageHelper.cs
