+++ Meio.App/Views/PlaylistView.axaml.cs
