+++ Meio.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meio.app", "Meio.app\Meio.app.csproj", "{A0DA8FBF-856F-44FE-89D5-C394F5F51E16}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Meio.App", "Meio.App\Meio.App.csproj", "{A0DA8FBF-856F-44FE-89D5-C394F5F51E16}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
